Rename marker algebra submodules
