+++ nab-python/src/nab_python/_marker_algebra/errors.py
