+++ nab-python/src/nab_python/_marker_algebra/markerset.py
@@ -5,15 +5,15 @@ from __future__ import annotations
 from functools import wraps
 from typing import TYPE_CHECKING, ParamSpec, TypeVar
 
-from . import _atoms, _engine
-from ._errors import ComplexityLimitExceeded, UnserializableSet
+from . import atoms, engine
+from .errors import ComplexityLimitExceeded, UnserializableSet
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Mapping
     from collections.abc import Set as AbstractSet
 
     from .._vendor.packaging.markers import Marker
-    from ._atoms import Formula
+    from .atoms import Formula
 
 DEFAULT_MAX_CELLS = 100_000
 
@@ -66,56 +66,56 @@ class MarkerSet:
     ) -> MarkerSet:
         """Build the set of a marker string or :class:`Marker`."""
         _check_max_cells(max_cells)
-        return cls(_atoms.parse(marker), max_cells)
+        return cls(atoms.parse(marker), max_cells)
 
     @classmethod
     def true(cls, *, max_cells: int = DEFAULT_MAX_CELLS) -> MarkerSet:
         """Return the universal set: every environment (an absent marker)."""
         _check_max_cells(max_cells)
-        return cls(_atoms.TRUE, max_cells)
+        return cls(atoms.TRUE, max_cells)
 
     @classmethod
     def false(cls, *, max_cells: int = DEFAULT_MAX_CELLS) -> MarkerSet:
         """Return the empty set: no environment."""
         _check_max_cells(max_cells)
-        return cls(_atoms.FALSE, max_cells)
+        return cls(atoms.FALSE, max_cells)
 
     # ---- algebra
 
     def __and__(self, other: MarkerSet) -> MarkerSet:
-        return MarkerSet(_atoms.make_and((self._tree, other._tree)), self._max_cells)
+        return MarkerSet(atoms.make_and((self._tree, other._tree)), self._max_cells)
 
     def __or__(self, other: MarkerSet) -> MarkerSet:
-        return MarkerSet(_atoms.make_or((self._tree, other._tree)), self._max_cells)
+        return MarkerSet(atoms.make_or((self._tree, other._tree)), self._max_cells)
 
     def complement(self) -> MarkerSet:
         """Return the set of environments this set excludes."""
-        return MarkerSet(_atoms.make_not(self._tree), self._max_cells)
+        return MarkerSet(atoms.make_not(self._tree), self._max_cells)
 
     # ---- decision procedures
 
     @_bounded
     def is_empty(self) -> bool:
         """Whether the set is unsatisfiable."""
-        return _engine.is_empty(self._tree, self._max_cells)
+        return engine.is_empty(self._tree, self._max_cells)
 
     @_bounded
     def is_tautology(self) -> bool:
         """Whether the set is every environment."""
-        return _engine.is_empty(_atoms.make_not(self._tree), self._max_cells)
+        return engine.is_empty(atoms.make_not(self._tree), self._max_cells)
 
     @_bounded
     def is_disjoint(self, other: MarkerSet) -> bool:
         """Whether the two sets share no environment."""
-        return _engine.is_empty(
-            _atoms.make_and((self._tree, other._tree)), self._max_cells
+        return engine.is_empty(
+            atoms.make_and((self._tree, other._tree)), self._max_cells
         )
 
     @_bounded
     def implies(self, other: MarkerSet) -> bool:
         """Whether every environment in this set is in ``other``."""
-        return _engine.is_empty(
-            _atoms.make_and((self._tree, _atoms.make_not(other._tree))),
+        return engine.is_empty(
+            atoms.make_and((self._tree, atoms.make_not(other._tree))),
             self._max_cells,
         )
 
@@ -146,33 +146,33 @@ class MarkerSet:
             )
             raise ValueError(msg)
         if on_unknown_variable == "error":
-            missing = _engine.unprovided_variables(self._tree, env)
+            missing = engine.unprovided_variables(self._tree, env)
             if missing:
                 msg = f"restrict() has no value for {sorted(missing)}"
                 raise ValueError(msg)
-        return MarkerSet(_engine.restrict_tree(self._tree, env), self._max_cells)
+        return MarkerSet(engine.restrict_tree(self._tree, env), self._max_cells)
 
     @_bounded
     def variables(self) -> frozenset[str]:
         """Return the variables the set references, as written."""
-        return _engine.variables_of(self._tree)
+        return engine.variables_of(self._tree)
 
     @_bounded
     def membership_literals(self) -> frozenset[tuple[str, str]]:
         """Return the ``(variable, canonical name)`` set-memberships the set tests."""
-        return _engine.membership_literals_of(self._tree)
+        return engine.membership_literals_of(self._tree)
 
     # ---- evaluation and witness
 
     @_bounded
     def evaluate(self, env: Mapping[str, str | AbstractSet[str]]) -> bool:
         """Whether a full environment is in the set (extras are sets)."""
-        return _engine.evaluate_tree(self._tree, env)
+        return engine.evaluate_tree(self._tree, env)
 
     @_bounded
     def witness(self) -> dict[str, str | frozenset[str]] | None:
         """Return a satisfying environment, or ``None`` when none is found."""
-        return _engine.witness(self._tree, self._max_cells)
+        return engine.witness(self._tree, self._max_cells)
 
     # ---- serialisation
 
@@ -190,7 +190,7 @@ class MarkerSet:
         if self.is_empty():
             msg = "the empty set has no marker string"
             raise UnserializableSet(msg)
-        text = _engine.serialize(_engine.to_nnf(self._tree))
+        text = engine.serialize(engine.to_nnf(self._tree))
         rebuilt = MarkerSet.from_marker(text, max_cells=self._max_cells)
         if not self.equivalent(rebuilt):  # pragma: no cover
             # A last-resort guard: the per-atom complements are sound by
